@@ -8,11 +8,15 @@
     "@mui/icons-material": "^5.6.1",
     "@mui/material": "^5.6.1",
     "@mui/styled-engine": "^5.6.1",
+    "@reduxjs/toolkit": "^1.8.1",
     "cra-template": "1.1.3",
-    "react": "^18.0.0",
+    "react": "^18.1.0",
     "react-dom": "^18.0.0",
     "react-glitch-effect": "^3.1.0",
-    "react-scripts": "5.0.0"
+    "react-redux": "^8.0.1",
+    "react-scripts": "^5.0.1",
+    "redux": "^4.2.0",
+    "redux-toolkit": "^1.1.2"
   },
   "scripts": {
     "start": "react-scripts start",
